requirements.txt
Flask==2.2.5
google-cloud-pubsub==2.21.0
gunicorn==21.2.0
